Migrate remaining lint/fix/fmt to call-by-name.
