+++ src/python/pants/core/goals/lint_test.py
@@ -23,8 +23,8 @@ from pants.core.goals.lint import (
     LintSubsystem,
     LintTargetsRequest,
     Partitions,
-    lint,
 )
+from pants.core.goals.lint_goal import lint
 from pants.core.util_rules.distdir import DistDir
 from pants.core.util_rules.partitions import PartitionerType, _EmptyMetadata
 from pants.engine.addresses import Address
@@ -36,7 +36,7 @@ from pants.engine.unions import UnionMembership, UnionRule
 from pants.option.option_types import SkipOption
 from pants.option.subsystem import Subsystem
 from pants.testutil.option_util import create_goal_subsystem
-from pants.testutil.rule_runner import MockGet, RuleRunner, mock_console, run_rule_with_mocks
+from pants.testutil.rule_runner import RuleRunner, mock_console, run_rule_with_mocks
 from pants.util.logging import LogLevel
 from pants.util.meta import classproperty
 
@@ -211,7 +211,9 @@ def mock_file_partitioner(__implicitly: dict) -> Partitions[str, Any]:
     return Partitions.single_partition(request.files)
 
 
-def mock_lint_partition(request: Any) -> LintResult:
+def mock_lint_partition(__implicitly: dict) -> LintResult:
+    request, typ = next(iter(__implicitly[0].items()))
+    assert typ == AbstractLintRequest.Batch
     request_type = {cls.Batch: cls for cls in _all_lint_requests()}[type(request)]
     return request_type.get_lint_result(request.elements)
 
@@ -308,6 +310,11 @@ class BuildFileFixer(MockLintRequest, FixFilesRequest):
         return LintResult(0, "", "", cls.tool_name)
 
 
+def mock_fix_partition_as_linter(request) -> LintResult:
+    request_type = {cls.Batch: cls for cls in _all_lint_requests()}[type(request)]
+    return request_type.get_lint_result(request.elements)
+
+
 @pytest.fixture
 def rule_runner() -> RuleRunner:
     return RuleRunner()
@@ -364,13 +371,6 @@ def run_lint_rule(
                 union_membership,
                 DistDir(relpath=Path("dist")),
             ],
-            mock_gets=[
-                MockGet(
-                    output_type=LintResult,
-                    input_types=(AbstractLintRequest.Batch,),
-                    mock=mock_lint_partition,
-                ),
-            ],
             mock_calls={
                 "pants.engine.internals.graph.filter_targets": lambda __implicitly: FilteredTargets(
                     tuple(targets)
@@ -380,6 +380,8 @@ def run_lint_rule(
                 ),
                 "pants.core.goals.lint.partition_targets": mock_target_partitioner,
                 "pants.core.goals.lint.partition_files": mock_file_partitioner,
+                "pants.core.goals.lint.lint_batch": mock_lint_partition,
+                "pants.core.goals.lint_goal.run_fixer_or_formatter_as_linter": mock_fix_partition_as_linter,
                 "pants.engine.intrinsics.digest_to_snapshot": lambda __implicitly: EMPTY_SNAPSHOT,
             },
             union_membership=union_membership,
