+++ src/python/pants/core/register.py
@@ -24,6 +24,7 @@ from pants.core.goals import (
     generate_lockfiles,
     generate_snapshots,
     lint,
+    lint_goal,
     package,
     publish,
     repl,
@@ -79,6 +80,7 @@ def rules():
         *generate_lockfiles.rules(),
         *generate_snapshots.rules(),
         *lint.rules(),
+        *lint_goal.rules(),
         *update_build_files.rules(),
         *package.rules(),
         *publish.rules(),
